% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-helpers.R
\name{check_grid_stack_item_list}
\alias{check_grid_stack_item_list}
\title{Checks that all given arguments are \code{grid_stack_item}s}
\usage{
check_grid_stack_item_list(...)
}
\arguments{
\item{...}{arguments to be checked}
}
\value{
TRUE if arguments are valid \code{grid_stack_item}s, FALSE otherwise
}
\description{
Checks that all given arguments are \code{grid_stack_item}s
}
