% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridstackeR_demo.R
\name{gridstackeR_demo}
\alias{gridstackeR_demo}
\title{Demo}
\usage{
gridstackeR_demo()
}
\value{
an example shiny shinyApp that uses the gridstackeR package to create
a responsive layout with resizable and draggable boxes.
}
\description{
a short example of gridstackeR
}
\examples{
\dontrun{
gridstackeR_demo()
}

}
